+++ controller/taskController.js
@@ -1,4 +1,4 @@
-mavikentApp.controller("TaskCtrl", function($scope, $state, $http, $localStorage, $rootScope) {
+mavikentApp.controller("TaskCtrl", function ($scope, $state, $http, $localStorage, $rootScope) {
 
     var token = $rootScope.mkb.token;
     $scope.IsEdit = false;
@@ -12,27 +12,27 @@ mavikentApp.controller("TaskCtrl", function($scope, $state, $http, $localStorage
         updated_by: $rootScope.mkb.current_user.name
     };
     //list all item
-    $http.get(host + "/api/task?token=" + token).success(function(resp) {
+    $http.get(host + "/api/task?token=" + token).success(function (resp) {
         if (resp.status == false) {
             console.log("error : ", JSON.stringify(resp));
-            stateControl(resp.code,resp.data);
+            stateControl(resp.code, resp.data);
             return;
         }
         $scope.list = resp.data;
 
 
-    }).error(function(err) {
+    }).error(function (err) {
         console.error(JSON.stringify(err));
         sweetAlert("Oops...", "Bir hata oluştu", "error");
     });
 
     //add function
-    $scope.save = function() {
+    $scope.save = function () {
         if ($scope.IsEdit) {
-            $http.put(host + "/api/task/?token=" + token, $scope.obj).success(function(resp) {
+            $http.put(host + "/api/task/?token=" + token, $scope.obj).success(function (resp) {
                 if (resp.status == false) {
                     console.log("error : ", JSON.stringify(resp));
-                    stateControl(resp.code,resp.data);
+                    stateControl(resp.code, resp.data);
                     return;
                 }
                 $scope.IsEdit = false;
@@ -44,16 +44,16 @@ mavikentApp.controller("TaskCtrl", function($scope, $state, $http, $localStorage
                     updated_by: $rootScope.mkb.current_user.name
                 };
                 swal("Başarılı!", "Güncelleme Başarılı!", "success")
-            }).error(function(err) {
+            }).error(function (err) {
                 console.error(JSON.stringify(err));
                 sweetAlert("Oops...", "Bir hata oluştu", "error");
             });
 
         } else {
-            $http.post(host + "/api/task?token=" + token, $scope.obj).success(function(resp) {
+            $http.post(host + "/api/task?token=" + token, $scope.obj).success(function (resp) {
                 if (resp.status == false) {
                     console.log("error : ", JSON.stringify(resp));
-                    stateControl(resp.code,resp.data);
+                    stateControl(resp.code, resp.data);
                     return;
                 }
                 $scope.list.push(resp.data);
@@ -64,7 +64,7 @@ mavikentApp.controller("TaskCtrl", function($scope, $state, $http, $localStorage
                 };
                 // console.log(JSON.stringify(resp)); 
                 swal("Başarılı!", "Ekleme Başarılı!", "success")
-            }).error(function(err) {
+            }).error(function (err) {
                 console.error(JSON.stringify(err));
                 sweetAlert("Oops...", "Bir hata oluştu", "error");
             });
@@ -74,34 +74,34 @@ mavikentApp.controller("TaskCtrl", function($scope, $state, $http, $localStorage
 
     //delete function
 
-    $scope.delete = function(id, index) {
-        swal({  
-            title: "Emin misiniz?",   
-            text: "Bu öğeyi silmek istedğinizden emin misiniz?",   
-            type: "warning",  
-            showCancelButton: true,   
-            confirmButtonColor: "#DD6B55",   
-            confirmButtonText: "Evet",   
-            closeOnConfirm: false }, function(){
-        $http.delete(host + "/api/task/" + id + "?token=" + token).success(function(resp) {
-            if (resp.status == false) {
-                console.log("error : ", JSON.stringify(resp));
-                stateControl(resp.code,resp.data);
-                return;
-            }
-            $scope.list.splice(index, 1);
-            console.log(JSON.stringify(resp));
-            swal("Başarılı!", "Silme Başarılı!", "success")
-        }).error(function(err) {
-            console.error(JSON.stringify(err));
-            sweetAlert("Oops...", "Bir hata oluştu", "error");
-        });
+    $scope.delete = function (id, index) {
+        swal({
+            title: "Emin misiniz?",
+            text: "Bu öğeyi silmek istedğinizden emin misiniz?",
+            type: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#DD6B55",
+            confirmButtonText: "Evet",
+            closeOnConfirm: false
+        }, function () {
+            $http.delete(host + "/api/task/" + id + "?token=" + token).success(function (resp) {
+                if (resp.status == false) {
+                    console.log("error : ", JSON.stringify(resp));
+                    stateControl(resp.code, resp.data);
+                    return;
+                }
+                $scope.list.splice(index, 1);
+                swal("Başarılı!", "Silme Başarılı!", "success")
+            }).error(function (err) {
+                console.error(JSON.stringify(err));
+                sweetAlert("Oops...", "Bir hata oluştu", "error");
+            });
         });
     }
 
     //edit function
 
-    $scope.edit = function(id, index) {
+    $scope.edit = function (id, index) {
         $scope.IsEdit = true;
         $scope.listIndex = index;
         $scope.editId = id;
